requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
ta==0.10.2
websocket-client==1.6.1
schedule==1.2.0
